requirements.txt
Flask==2.3.3
GitPython==3.1.40
langchain==0.0.350
openai==1.3.7
python-dotenv==1.0.0
requests==2.31.0
markdown==3.5.1
